+++ src/convert_roms/test_chd_parents_to_chd_should_not_touch_parent.rs
@@ -92,6 +92,7 @@ async fn test() {
         &[],
         &None,
         true,
+        false,
     )
     .await
     .unwrap();
@@ -133,6 +134,7 @@ async fn test() {
         &[],
         &None,
         true,
+        false,
     )
     .await
     .unwrap();