+++ src/convert_roms/test_chd_parents_to_iso.rs
@@ -92,6 +92,7 @@ async fn test() {
         &[],
         &None,
         true,
+        false,
     )
     .await
     .unwrap();