+++ src/convert_roms/test_iso_chd_to_chd.rs
@@ -88,6 +88,7 @@ async fn test() {
         &[],
         &None,
         false,
+        false,
     )
     .await
     .unwrap();