+++ src/convert_roms/test_iso_to_chd_parents.rs
@@ -93,6 +93,7 @@ async fn test() {
         &[],
         &None,
         true,
+        false,
     )
     .await
     .unwrap();