+++ src/convert_roms/test_multiple_tracks_chd_to_chd.rs
@@ -97,6 +97,7 @@ async fn test() {
         &[],
         &None,
         false,
+        false,
     )
     .await
     .unwrap();