+++ src/convert_roms/test_sevenzip_iso_to_chd.rs
@@ -82,6 +82,7 @@ async fn test() {
         &[],
         &None,
         false,
+        false,
     )
     .await
     .unwrap();