+++ src/convert_roms/test_sevenzip_multiple_tracks_cue_bin_to_chd.rs
@@ -75,6 +75,7 @@ async fn test() {
         &[],
         &None,
         false,
+        false,
     )
     .await
     .unwrap();