+++ src/prompt.rs
@@ -3,6 +3,7 @@ use super::model::*;
 use dialoguer::{Confirm, MultiSelect, Select};
 use simple_error::SimpleResult;
 use sqlx::sqlite::SqliteConnection;
+use std::path::PathBuf;
 
 pub async fn prompt_for_systems(
     connection: &mut SqliteConnection,
@@ -143,6 +144,36 @@ pub fn prompt_for_rom_game_system(
         Some(_) => index.map(|i| roms_games_systems.remove(i - 1)),
         None => None,
     })
+}
+
+pub async fn prompt_for_parent_romfile(
+    connection: &mut SqliteConnection,
+    system_id: i64,
+    extension: &str,
+) -> SimpleResult<Option<Romfile>> {
+    let mut romfiles =
+        find_romfiles_by_system_id_and_extension_and_no_parent_id(connection, system_id, extension)
+            .await;
+    let index = match romfiles.len() {
+        0 => None,
+        _ => select_opt(
+            &romfiles
+                .iter()
+                .map(|romfile| {
+                    PathBuf::from(&romfile.path)
+                        .file_name()
+                        .unwrap()
+                        .to_str()
+                        .unwrap()
+                        .to_string()
+                })
+                .collect::<Vec<String>>(),
+            "Please select a ROM file",
+            None,
+            None,
+        )?,
+    };
+    Ok(index.map(|index| romfiles.remove(index)))
 }
 
 pub fn confirm(default: bool) -> SimpleResult<bool> {