+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add configuration options for chdman
 - Add support for CHD parents, enable via the `CHD_PARENTS` settings
 - Add a `-u` flag to `config` to unset nullable settings
+- Add a `-p` flag to `convert-roms` to prompt for an optional CHD parent when converting to CHD 
 
 # 0.18.1
 - Bump minimum required chdman version to 0.264 for dreamcast